handle distinct / duplicate on the frontend,
