+++ frontend/src/utils/en.json
@@ -166,7 +166,7 @@
   "Are you sure you want to mark these records as duplicates? They will be removed from Golden Records upon ticket closure.": "Are you sure you want to mark these records as duplicates? They will be removed from Golden Records upon ticket closure.",
   "Create Linked Ticket": "Create Linked Ticket",
   "Confirmation": "Confirmation",
-  "Mark Duplicate": "Mark Duplicate",
+  "Mark as Duplicate": "Mark as Duplicate",
   "Similarity Score": "Similarity Score",
   "Last Registration Date": "Last Registration Date",
   "Doc Type": "Doc Type",
@@ -876,5 +876,8 @@
   "Authorized by": "Authorized by",
   "This individual is also included in other Payment Plans. Click this icon to view details.": "This individual is also included in other Payment Plans. Click this icon to view details.",
   "AdditionalRegistrationInformation": "Additional Registration Information",
-  "Delivery Mechanism Data to be edited": "Delivery Mechanism Data to be edited"
+  "Delivery Mechanism Data to be edited": "Delivery Mechanism Data to be edited",
+  "Mark as Distinct": "Mark as Distinct",
+  "Uniqueness": "Uniqueness",
+  "Are you sure you want to clear the selected ids? They won't be marked as a duplicate or distinct anymore.": "Are you sure you want to clear the selected ids? They won't be marked as a duplicate or distinct anymore."
 }
